Extract schema utilities and dispatch primitives from mcp.tools
First of three planned PRs decomposing backend/mcp/tools.py (4910 LOC).
This pass extracts the lowest-layer primitives — no behavior change,
pure mechanical movement.

Converts the tools module into a package:
- tools/__init__.py — the rest (handlers, output schemas, registry, public API)
- tools/constants.py — shared enum-value tuples (filter fields, count modes, etc.)
- tools/schema_utils.py — JSON-Schema builders (_object_schema, _array_of,
  _schema_input_schema, _field_json_schema, _filter_option_metadata, etc.)
- tools/args_schemas.py — Marshmallow Schema subclasses for MCP tool args
- tools/dispatch.py — McpToolSpec / McpToolResult / McpOutputValidationError
  plus _validate_payload / _require_scope / output validation walker

External imports of backend.mcp.tools (call_tool, tool_definitions,
McpToolResult, McpOutputValidationError, _tool_specs,
_server_capabilities_payload) keep working — those names remain at the
package top level via the new submodule re-imports.

PRs B and C will extract output schemas and per-tool handlers next.
